move the GeneratorConfig to it's own directory
since it's also used by the analyzer
