+++ tools/serverpod_cli/lib/src/analyzer/dart/endpoints_analyzer.dart
@@ -10,7 +10,7 @@ import 'package:analyzer/file_system/physical_file_system.dart';
 import 'package:source_span/source_span.dart';
 
 import '../../util/subdirectory_extraction.dart';
-import '../../generator/config.dart';
+import '../../config/config.dart';
 import 'definitions.dart';
 import '../code_analysis_collector.dart';
 import '../../generator/types.dart';
